Move copied code into their own header files.
This more closely matches the structure they were copied
from, and allows us to more easily keep their old licenses.
